requirements.txt
streamlit==1.33.0
gitpython==3.1.43
transformers==4.41.2
torch==2.3.0
plotly==5.22.0
pandas==2.2.2
networkx==3.3